requirements.txt
# API Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.1

# SurrealDB
httpx==0.27.2

# Embeddings
openai==1.54.3
sentence-transformers==3.3.1
torch==2.5.1

# Email Processing
email-validator==2.2.0
python-dateutil==2.9.0.post0

# NLP & Text Processing
langdetect==1.0.9
spacy==3.8.2

# Utilities
python-dotenv==1.0.1
